extract out and organize diff shapes to simplify this file a bit
