+++ app/src/models/diff/raw-diff.ts
@@ -1,0 +1,87 @@
+import { DiffLine } from './diff-line'
+
+/** each diff is made up of a number of hunks */
+export class DiffHunk {
+  /** details from the diff hunk header about the line start and patch length */
+  public readonly header: DiffHunkHeader
+  /** the contents - context and changes - of the diff setion */
+  public readonly lines: ReadonlyArray<DiffLine>
+  /** the diff hunk's start position in the overall file diff */
+  public readonly unifiedDiffStart: number
+  /** the diff hunk's end position in the overall file diff */
+  public readonly unifiedDiffEnd: number
+
+  public constructor(
+    header: DiffHunkHeader,
+    lines: ReadonlyArray<DiffLine>,
+    unifiedDiffStart: number,
+    unifiedDiffEnd: number
+  ) {
+    this.header = header
+    this.unifiedDiffStart = unifiedDiffStart
+    this.unifiedDiffEnd = unifiedDiffEnd
+    this.lines = lines
+  }
+}
+
+/** details about the start and end of a diff hunk */
+export class DiffHunkHeader {
+  /** The line in the old (or original) file where this diff hunk starts */
+  public readonly oldStartLine: number
+
+  /** The number of lines in the old (or original) file that this diff hunk covers */
+  public readonly oldLineCount: number
+
+  /** The line in the new file where this diff hunk starts */
+  public readonly newStartLine: number
+
+  /** The number of lines in the new file that this diff hunk covers */
+  public readonly newLineCount: number
+
+  public constructor(
+    oldStartLine: number,
+    oldLineCount: number,
+    newStartLine: number,
+    newLineCount: number,
+    sectionHeading?: string | null
+  ) {
+    this.oldStartLine = oldStartLine
+    this.oldLineCount = oldLineCount
+    this.newStartLine = newStartLine
+    this.newLineCount = newLineCount
+  }
+}
+
+/** the contents of a diff generated by Git */
+export interface IRawDiff {
+  /**
+   * The plain text contents of the diff header. This contains
+   * everything from the start of the diff up until the first
+   * hunk header starts. Note that this does not include a trailing
+   * newline.
+   */
+  readonly header: string
+
+  /**
+   * The plain text contents of the diff. This contains everything
+   * after the diff header until the last character in the diff.
+   *
+   * Note that this does not include a trailing newline nor does
+   * it include diff 'no newline at end of file' comments. For
+   * no-newline information, consult the DiffLine noTrailingNewLine
+   * property.
+   */
+  readonly contents: string
+
+  /**
+   * Each hunk in the diff with information about start, and end
+   * positions, lines and line statuses.
+   */
+  readonly hunks: ReadonlyArray<DiffHunk>
+
+  /**
+   * Whether or not the unified diff indicates that the contents
+   * could not be diffed due to one of the versions being binary.
+   */
+  readonly isBinary: boolean
+}
